Changed sdp.packet_duration to be a float Added example of Livewire SDP file found on twitter: https://twitter.com/verySetya/status/1079900945725644800

+++ src/mast.h
@@ -146,7 +146,7 @@ typedef struct
     int encoding;
     int sample_rate;
     int channel_count;
-    int packet_duration;
+    float packet_duration;
 
 } mast_sdp_t;
